Noninteractive mode and reading input from pipe
Noninteractive mode is started if any input is piped to the
program. In this mode computer prints to 'stdout', there is no
user interface, 'usleep' is not called between cycles, and when
program reads from last address, instead of random value, it
gets one line from stdin (pipe). In this line every '*' is
interrupted as true and all other characters as false. If there
is not enough characters, all missing bits are set to false; and
all characters after the eight one are ignored.